refactor(orchestrator): extract ContextualToolsProvider
R2 step 2 final provider: moves addUserMemoryToolIfEnabled /
addConversationRecallToolIfEnabled / addReadAttachmentToolIfEnabled into
one provider rather than the three the plan lists - all three are enabled
by what the conversation holds rather than by the whitelist, and their
per-tool provenance tags come from ToolObjectReflector (not the provider's
source()), so memory:*/recall:* approval patterns are unaffected.

Per-call construction again: attachmentStore/attachmentTextExtractor are
field-injected and null at constructor time.

AgentOrchestrator: 2,108 -> 2,031 lines (from 2,725 at R2 start). All 8
tool sources are now SPI-conformant provider classes.
